requirements.txt
fastapi[all]
peewee
pydantic
python-dotenv
psycopg2-binary
PyJWT